+++ GameServerLib/Items/Shop.cs
@@ -78,7 +78,7 @@ namespace LeagueSandbox.GameServer.Items
             var inventory = _owner.Inventory;
             _owner.Stats.RemoveModifier(item.ItemType);
             _game.PacketNotifier.NotifyRemoveItem(_owner, slotId, stackSize);
-            _owner.RemoveSpell(slotId + ITEM_ACTIVE_OFFSET);
+            _owner.RemoveSpell((byte)(slotId + ITEM_ACTIVE_OFFSET));
             inventory.RemoveItem(item);
         }
 
@@ -93,7 +93,7 @@ namespace LeagueSandbox.GameServer.Items
             _game.PacketNotifier.NotifyItemBought(_owner, i);
             if (!string.IsNullOrEmpty(i.ItemType.SpellName))
             {
-                _owner.SetSpell(i.ItemType.SpellName, _owner.Inventory.GetItemSlot(i) + ITEM_ACTIVE_OFFSET, true);
+                _owner.SetSpell(i.ItemType.SpellName, (byte)(_owner.Inventory.GetItemSlot(i) + ITEM_ACTIVE_OFFSET), true);
             }
             return true;
         }
